hotfix(Dockerfile): fix permissions issues and correct PATH
Since for security and good practices reasons we are using the `devops`
user instead of `root`, pip is installing the libraries binaries at
`$HOME/.local/bin`. The problem is that by default that directory is not
added to PATH by default, so we need to set it **before** using any of
the python packages installed through pip.

Also, Django needs the path where SQLite file is installed (`/app`) to
be writable by the `devops` user. So we are creating and changing the
permissions of that directory.

This commit fix both issues.